refactor: CSP 설정 및 App.tsx 컴포넌트/훅 분리 (#18, #19)
- CSP를 null에서 'default-src self; style-src self unsafe-inline'로 변경
- StackItem 타입을 src/types.ts로 추출
- useStore, useCommands 커스텀 훅 분리
- InputBar, StackItem, EmptyState 컴포넌트 분리
- App.tsx를 351줄에서 212줄로 축소
