+++ src/components/EmptyState.tsx
@@ -1,0 +1,9 @@
+export function EmptyState() {
+  return (
+    <div className="empty-state">
+      /del # · /edit # · /pop · /clear
+      <br />
+      Ctrl+Shift+T Focus · Ctrl+Shift+P Pin
+    </div>
+  );
+}
