+++ src/types.ts
@@ -1,0 +1,4 @@
+export interface StackItem {
+  id: number;
+  text: string;
+}
